feat(vm:inspect_as_batch): split tx internally
