+++ crates/zksync/core/src/vm/mod.rs
@@ -1,13 +1,11 @@
 mod db;
 mod env;
-mod factory_deps;
 mod farcall;
 mod inspect;
 mod runner;
 mod storage_view;
 mod tracer;
 
-pub use factory_deps::split_tx_by_factory_deps;
-pub use inspect::{inspect, inspect_multi, ZKVMExecutionResult, ZKVMResult};
+pub use inspect::{inspect, inspect_as_batch, ZKVMExecutionResult, ZKVMResult};
 pub use runner::{balance, call, code_hash, create, encode_create_params, nonce, transact};
 pub use tracer::CheatcodeTracerContext;
